feat(jobProfile): Can edit job profiles

+++ api/src/services/jobProfiles/jobProfiles.ts
@@ -5,11 +5,12 @@ import type {
 } from 'types/graphql'
 
 import { validate } from '@redwoodjs/api'
+import { SyntaxError } from '@redwoodjs/graphql-server'
 
 import { db } from 'src/lib/db'
 
 export const jobProfiles: QueryResolvers['jobProfiles'] = () => {
-  return db.jobProfile.findMany()
+  return db.jobProfile.findMany({ orderBy: { createdAt: 'desc' } })
 }
 
 export const jobProfile: QueryResolvers['jobProfile'] = ({ id }) => {
@@ -54,9 +55,21 @@ export const updateJobProfile: MutationResolvers['updateJobProfile'] = ({
   })
 }
 
-export const deleteJobProfile: MutationResolvers['deleteJobProfile'] = ({
+export const deleteJobProfile: MutationResolvers['deleteJobProfile'] = async ({
   id,
 }) => {
+  const linkedWorkRequests = await db.workRequest.findMany({
+    where: {
+      jobProfileId: id,
+    },
+  })
+
+  if (linkedWorkRequests.length > 0) {
+    throw new SyntaxError(
+      'Er zijn werkaanvragen gekoppeld aan dit functieprofiel. Je kunt alleen functieprofielen zonder werkaanvragen verwijderen.'
+    )
+  }
+
   return db.jobProfile.delete({
     where: { id },
   })
